game board, charechter, roll dice, (pairing with Marek)
