+++ FEMO/GameScene.swift
@@ -1,9 +1,9 @@
 //
 //  GameScene.swift
-//  FEMO
+//  Indiana Codes
 //
-//  Created by Farah Jabri on 19/09/2018.
-//  Copyright © 2018 FEMO@Makers. All rights reserved.
+//  Created by Marek Schmidt on 9/19/18.
+//  Copyright © 2018 Marek Schmidt. All rights reserved.
 //
 
 import SpriteKit
@@ -11,75 +11,90 @@ import GameplayKit
 
 class GameScene: SKScene {
     
-    private var label : SKLabelNode?
-    private var spinnyNode : SKShapeNode?
+    var tilesArray:[SKSpriteNode]? = [SKSpriteNode]()
+    var player1:SKSpriteNode?
+    
+    var currentTile = 0
+    var movingToTile = false
+    
+    var dieRoll = 0
+    
+    let moveSound = SKAction.playSoundFileNamed("tap.wav", waitForCompletion: false)
+    
+    func setupTiles() {
+        for i in 1...21 {
+            if let tile = self.childNode(withName: "tile\(i)") as? SKSpriteNode {
+                tilesArray?.append(tile)
+            }
+        }
+    }
+    
+    func createPlayer1() {
+        player1 = SKSpriteNode(imageNamed: "character")
+        
+        guard let player1PositionX = tilesArray?.first?.position.x else {return}
+        guard let player1PositionY = tilesArray?.first?.position.y else {return}
+        player1?.position = CGPoint(x: player1PositionX, y: player1PositionY + 10)
+        
+        self.addChild(player1!)
+    }
     
     override func didMove(to view: SKView) {
+        setupTiles()
+        createPlayer1()
+    }
+    
+    func moveToNextTile() {
+        player1?.removeAllActions()
+        movingToTile = true
         
-        // Get label node from scene and store it for use later
-        self.label = self.childNode(withName: "//helloLabel") as? SKLabelNode
-        if let label = self.label {
-            label.alpha = 0.0
-            label.run(SKAction.fadeIn(withDuration: 2.0))
-        }
+        guard let nextTile = tilesArray?[currentTile + 1].position else {return}
         
-        // Create shape node to use during mouse interaction
-        let w = (self.size.width + self.size.height) * 0.05
-        self.spinnyNode = SKShapeNode.init(rectOf: CGSize.init(width: w, height: w), cornerRadius: w * 0.3)
-        
-        if let spinnyNode = self.spinnyNode {
-            spinnyNode.lineWidth = 2.5
+        if let player1 = self.player1 {
+            let moveAction = SKAction.move(to: CGPoint(x: nextTile.x, y: nextTile.y + 10), duration: 0.8)
+            player1.run(moveAction, completion: {
+                self.movingToTile = false
+            })
+            currentTile += 1
             
-            spinnyNode.run(SKAction.repeatForever(SKAction.rotate(byAngle: CGFloat(Double.pi), duration: 1)))
-            spinnyNode.run(SKAction.sequence([SKAction.wait(forDuration: 0.5),
-                                              SKAction.fadeOut(withDuration: 0.5),
-                                              SKAction.removeFromParent()]))
+            self.run(moveSound)
         }
     }
     
-    
-    func touchDown(atPoint pos : CGPoint) {
-        if let n = self.spinnyNode?.copy() as! SKShapeNode? {
-            n.position = pos
-            n.strokeColor = SKColor.green
-            self.addChild(n)
-        }
+    func rollDie() {
+        let roll = arc4random_uniform(_:6) + 1
+        dieRoll = Int(roll)
     }
     
-    func touchMoved(toPoint pos : CGPoint) {
-        if let n = self.spinnyNode?.copy() as! SKShapeNode? {
-            n.position = pos
-            n.strokeColor = SKColor.blue
-            self.addChild(n)
-        }
-    }
-    
-    func touchUp(atPoint pos : CGPoint) {
-        if let n = self.spinnyNode?.copy() as! SKShapeNode? {
-            n.position = pos
-            n.strokeColor = SKColor.red
-            self.addChild(n)
+    func playTurn() {
+        rollDie()
+        print("You rolled \(dieRoll)")
+        for _ in 1 ... dieRoll {
+            moveToNextTile()
         }
     }
     
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
-        if let label = self.label {
-            label.run(SKAction.init(named: "Pulse")!, withKey: "fadeInOut")
+        if let touch = touches.first {
+            let location = touch.previousLocation(in: self)
+            let node = self.nodes(at: location).first
+            
+            if node?.name == "nextTileButton" {
+                playTurn()
+            }
         }
-        
-        for t in touches { self.touchDown(atPoint: t.location(in: self)) }
-    }
-    
-    override func touchesMoved(_ touches: Set<UITouch>, with event: UIEvent?) {
-        for t in touches { self.touchMoved(toPoint: t.location(in: self)) }
     }
     
     override func touchesEnded(_ touches: Set<UITouch>, with event: UIEvent?) {
-        for t in touches { self.touchUp(atPoint: t.location(in: self)) }
+        if !movingToTile {
+            player1?.removeAllActions()
+        }
     }
     
     override func touchesCancelled(_ touches: Set<UITouch>, with event: UIEvent?) {
-        for t in touches { self.touchUp(atPoint: t.location(in: self)) }
+        if !movingToTile {
+            player1?.removeAllActions()
+        }
     }
