--- Indiana Codes/GameScene.swift
@@ -1,104 +1,0 @@
-//
-//  GameScene.swift
-//  Indiana Codes
-//
-//  Created by Marek Schmidt on 9/19/18.
-//  Copyright © 2018 Marek Schmidt. All rights reserved.
-//
-
-import SpriteKit
-import GameplayKit
-
-class GameScene: SKScene {
-    
-    var tilesArray:[SKSpriteNode]? = [SKSpriteNode]()
-    var player1:SKSpriteNode?
-    
-    var currentTile = 0
-    var movingToTile = false
-    
-    var dieRoll = 0
-    
-    let moveSound = SKAction.playSoundFileNamed("tap.wav", waitForCompletion: false)
-    
-    func setupTiles() {
-        for i in 1...21 {
-            if let tile = self.childNode(withName: "tile\(i)") as? SKSpriteNode {
-                tilesArray?.append(tile)
-            }
-        }
-    }
-    
-    func createPlayer1() {
-        player1 = SKSpriteNode(imageNamed: "character")
-        
-        guard let player1PositionX = tilesArray?.first?.position.x else {return}
-        guard let player1PositionY = tilesArray?.first?.position.y else {return}
-        player1?.position = CGPoint(x: player1PositionX, y: player1PositionY + 10)
-        
-        self.addChild(player1!)
-    }
-    
-    override func didMove(to view: SKView) {
-        setupTiles()
-        createPlayer1()
-    }
-    
-    func moveToNextTile() {
-        player1?.removeAllActions()
-        movingToTile = true
-        
-        guard let nextTile = tilesArray?[currentTile + 1].position else {return}
-        
-        if let player1 = self.player1 {
-            let moveAction = SKAction.move(to: CGPoint(x: nextTile.x, y: nextTile.y + 10), duration: 0.8)
-            player1.run(moveAction, completion: {
-                self.movingToTile = false
-            })
-            currentTile += 1
-            
-            self.run(moveSound)
-        }
-    }
-    
-    func rollDie() {
-        let roll = Int.random(in: 1 ... 6)
-        dieRoll = roll
-    }
-    
-    func playTurn() {
-        rollDie()
-        print("You rolled \(dieRoll)")
-        for _ in 1 ... dieRoll {
-            moveToNextTile()
-        }
-    }
-    
-    override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
-        if let touch = touches.first {
-            let location = touch.previousLocation(in: self)
-            let node = self.nodes(at: location).first
-            
-            if node?.name == "nextTileButton" {
-                playTurn()
-            }
-        }
-    }
-    
-    override func touchesEnded(_ touches: Set<UITouch>, with event: UIEvent?) {
-        if !movingToTile {
-            player1?.removeAllActions()
-        }
-    }
-    
-    override func touchesCancelled(_ touches: Set<UITouch>, with event: UIEvent?) {
-        if !movingToTile {
-            player1?.removeAllActions()
-        }
-    }
-    
-    
-    override func update(_ currentTime: TimeInterval) {
-        // Called before each frame is rendered
-    }
-}
